+++ src/databases/NETCDF/avtFVCOMReader.C
@@ -3247,8 +3247,8 @@ avtFVCOMReader::GetVar(int timestate, const char *Variable, avtVariableCache *ca
           VarDimIDs[0] == nTimeID)
         {
           ntuples = dimSizes[VarDimIDs[1]];    
-          size_t starts[]={timestate,0};
-          size_t counts[]={1, ntuples};
+          size_t starts[]={static_cast<size_t>(timestate),0};
+          size_t counts[]={1, static_cast<size_t>(ntuples)};
           ptrdiff_t stride[]={1,1};
           float *vals = new float[ntuples];
 
@@ -3274,8 +3274,10 @@ avtFVCOMReader::GetVar(int timestate, const char *Variable, avtVariableCache *ca
           VarDimIDs[0] == nTimeID)
         {
           ntuples = dimSizes[VarDimIDs[1]]*dimSizes[VarDimIDs[2]];    
-          size_t starts[]={timestate,0,0};
-          size_t counts[]={1, dimSizes[VarDimIDs[1]], dimSizes[VarDimIDs[2]]};
+          size_t starts[]={static_cast<size_t>(timestate),0,0};
+          size_t counts[]={1,
+                           static_cast<size_t>(dimSizes[VarDimIDs[1]]),
+                           static_cast<size_t>(dimSizes[VarDimIDs[2]])};
           ptrdiff_t stride[]={1,1,1};
           float *vals = new float[ntuples];
